Ai Function Helper/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
python-dotenv
pocket-tts
scipy
pytest
httpx
